class to load multile files
