+++ requirements.txt
@@ -1,3 +1,7 @@
 pydantic
 requests
 pre-commit
+ecmwflibs
+eccodes
+cfgrib
+xarray
