+++ tests/test_orders.py
@@ -1,7 +1,7 @@
 import tempfile
 from unittest import mock
 
-from metofficeamd.app import MetOfficeAMD
+from metofficeamd.base import BaseMetOfficeAMD
 from tests.conftest import (
     mock_get_example_grib,
     mock_get_file_details,
@@ -11,14 +11,14 @@ from tests.conftest import (
 
 
 def test_init():
-    _ = MetOfficeAMD(client_id="fake", client_secret="fake")
+    _ = BaseMetOfficeAMD(client_id="fake", client_secret="fake")
 
 
 @mock.patch("requests.get")
 def test_get_orders(mock_get):
     mock_get.return_value = mock_get_order_list()
 
-    amd = MetOfficeAMD(client_id="fake", client_secret="fake")
+    amd = BaseMetOfficeAMD(client_id="fake", client_secret="fake")
     amd.get_orders()
 
 
@@ -28,7 +28,7 @@ def test_latest_order(mock_get):
 
     order_id = "test_order_id"
 
-    amd = MetOfficeAMD(client_id="fake", client_secret="fake")
+    amd = BaseMetOfficeAMD(client_id="fake", client_secret="fake")
     amd.get_lastest_order(order_id=order_id)
 
 
@@ -39,8 +39,8 @@ def test_latest_order_file_id(mock_get):
     order_id = "test_order_id"
     file_id = "atmosphere_high-cloud-cover+low-cloud-cover+medium-cloud-cover_+06_0"
 
-    amd = MetOfficeAMD(client_id="fake", client_secret="fake")
-    amd.get_lastest_order_file_id(order_id=order_id, file_id=file_id)
+    amd = BaseMetOfficeAMD(client_id="fake", client_secret="fake")
+    amd.get_latest_order_file_id(order_id=order_id, file_id=file_id)
 
 
 @mock.patch("requests.get")
@@ -50,5 +50,5 @@ def test_latest_order_file_id_data(mock_get):
     file_id = "atmosphere_high-cloud-cover+low-cloud-cover+medium-cloud-cover_+06_0"
 
     with tempfile.TemporaryDirectory() as tmpdirname:
-        amd = MetOfficeAMD(cache_dir=tmpdirname, client_id="fake", client_secret="fake")
+        amd = BaseMetOfficeAMD(cache_dir=tmpdirname, client_id="fake", client_secret="fake")
         amd.get_lastest_order_file_id_data(order_id=order_id, file_id=file_id)
