+++ tests/test_runs.py
@@ -1,7 +1,7 @@
 import tempfile
 from unittest import mock
 
-from metofficeamd.app import MetOfficeAMD
+from metofficeamd.base import BaseMetOfficeAMD
 from tests.conftest import mock_get_run_list, mock_get_run_list_for_model
 
 
@@ -9,7 +9,7 @@ from tests.conftest import mock_get_run_list, mock_get_run_list_for_model
 def test_get_runs(mock_get):
     mock_get.return_value = mock_get_run_list()
 
-    amd = MetOfficeAMD(client_id="fake", client_secret="fake")
+    amd = BaseMetOfficeAMD(client_id="fake", client_secret="fake")
     amd.get_runs()
 
 
@@ -19,5 +19,5 @@ def test_get_runs_model_id(mock_get):
 
     model_id = "mo-uk"
 
-    amd = MetOfficeAMD(client_id="fake", client_secret="fake")
+    amd = BaseMetOfficeAMD(client_id="fake", client_secret="fake")
     amd.get_runs_model_id(model_id=model_id)
